+++ cmd/generate.go
@@ -3,12 +3,9 @@ package cmd
 import (
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 
 	"github.com/fatih/color"
-	"github.com/samber/lo"
 	"github.com/spf13/cobra"
 	"github.com/teamkeel/keel/functions"
 )
@@ -21,30 +18,33 @@ var generateCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		schemaDir, _ := cmd.Flags().GetString("dir")
 
-		missing, err := keelDependenciesFulfilled(schemaDir)
+		packageJson, err := functions.NewPackageJson(filepath.Join(schemaDir, "package.json"))
 
 		if err != nil {
-			fmt.Println("⛔️ There is an issue with the package.json")
+			fmt.Println("⛔️ Could not create package.json automatically")
+			return
 		}
 
-		if len(missing) > 0 {
-			fmt.Println("In order to generate code, you need to install the following npm packages:")
+		err = packageJson.Bootstrap()
 
-			for _, item := range missing {
-				fmt.Printf("- %s\n", color.New(color.FgCyan).Sprint(item))
-			}
-
-			fmt.Println("Run the following command:")
-
-			packageList := strings.Join(missing, " ")
-			fmt.Println(color.New(color.FgGreen).Sprintf("npm install %s --save-dev", packageList))
+		if err != nil {
+			fmt.Println("⛔️ Could not bootstrap package.json")
 			return
 		}
 
 		r, err := functions.NewRuntime(schemaDir)
 
 		if err != nil {
-			fmt.Println("⛔️ Internal runtime error")
+			fmt.Println("⛔️ Internal runtime error (a)")
+			fmt.Print(err)
+			return
+		}
+
+		err = r.GenerateClientPackageJson()
+
+		if err != nil {
+			fmt.Println("⛔️ Could not generate @teamkeel/client package.json")
+			fmt.Println(err)
 
 			return
 		}
@@ -52,38 +52,58 @@ var generateCmd = &cobra.Command{
 		err = r.GenerateClient()
 
 		if err != nil {
-			fmt.Println("⛔️ Internal runtime error")
-
+			fmt.Println("⛔️ Internal runtime error (b)")
+			fmt.Print(err)
 			return
 		}
 
 		err = r.GenerateHandler()
 
 		if err != nil {
-			fmt.Println("⛔️ Internal runtime error")
-
+			fmt.Println("⛔️ Internal runtime error (c)")
+			fmt.Print(err)
 			return
+		}
+
+		sr, err := r.Scaffold()
+
+		if err != nil {
+			fmt.Println("⛔️ Internal runtime error (c)")
+			fmt.Print(err)
+			return
+		}
+		fmt.Printf("Generated the following files:\n\n")
+
+		fmt.Printf("--- %s ---\n", color.New(color.FgHiYellow).Sprint("Functions"))
+
+		if len(sr.CreatedFunctions) == 0 && sr.FunctionsCount > 0 {
+			fmt.Println("✅  No new functions to generate")
+		} else if sr.FunctionsCount == 0 {
+			fmt.Println("✅  No custom functions defined")
+		} else {
+			for _, f := range sr.CreatedFunctions {
+				fileName := filepath.Base(f)
+				fmt.Printf("⚡️ Generated %s %s\n", color.New(color.FgCyan).Sprint(fileName), color.New(color.Faint).Sprintf("[%s]", f))
+			}
 		}
 
 		result, errs := r.Bundle(true)
 
 		if len(errs) > 0 {
-			fmt.Println("⛔️ Internal runtime error")
+			fmt.Println("⛔️ Internal runtime error (d)")
+			fmt.Println(errs)
 		}
 
-		fmt.Println("🔨 Generating code...")
-
-		fmt.Println("---")
+		fmt.Printf("\n--- %s ---\n", color.New(color.FgHiGreen).Sprint("Runtime"))
 
 		for _, f := range result.OutputFiles {
 			lastFragment := filepath.Base(f.Path)
 
-			fmt.Printf("⚡️ Generated %s [%s]\n", lastFragment, color.New(color.FgCyan).Sprint(f.Path))
+			fmt.Printf("⚡️ Updated %s %s\n", color.New(color.FgCyan).Sprint(lastFragment), color.New(color.Faint).Sprintf("[%s]", f.Path))
 		}
 
 		fmt.Println("---")
 
-		fmt.Println("✅ Generation complete")
 	},
 }
 
@@ -102,38 +122,4 @@ func init() {
 var REQUIRED_DEPS = []string{
 	"@teamkeel/runtime",
 	"@teamkeel/sdk",
-}
-
-func keelDependenciesFulfilled(schemaDir string) (missing []string, err error) {
-	list := exec.Command("npm", "ls", "--parseable")
-	list.Dir = schemaDir
-
-	exhausted := REQUIRED_DEPS
-
-	output, err := list.CombinedOutput()
-	if err != nil {
-		return []string{}, err
-	}
-
-	outputStr := string(output)
-
-	splitStr := strings.Split(outputStr, "\n")
-
-	for _, s := range splitStr {
-		parts := strings.Split(s, "/")
-
-		if len(parts) >= 2 {
-			interestingParts := parts[len(parts)-2:]
-
-			packageName := strings.Join(interestingParts, "/")
-
-			if lo.Contains(exhausted, packageName) {
-				i := lo.IndexOf(exhausted, packageName)
-
-				exhausted = append(exhausted[:i], exhausted[i+1:]...)
-			}
-		}
-	}
-
-	return exhausted, nil
 }