+++ examples/Todo/functions/createAccount.ts
@@ -1,0 +1,7 @@
+import { CreateAccountInput, API } from '@teamkeel/sdk'
+
+export default async (inputs: CreateAccountInput, api: API) => {
+  // Build your universe
+
+  api.models.Account.create({ name: 'my account' })
+}