+++ functions/runtime.go
@@ -9,7 +9,10 @@ import (
 	"path"
 	"path/filepath"
 
+	_ "embed"
+
 	"github.com/evanw/esbuild/pkg/api"
+	"github.com/samber/lo"
 	"github.com/teamkeel/keel/proto"
 	"github.com/teamkeel/keel/schema"
 )
@@ -18,6 +21,17 @@ type Runtime struct {
 	Schema     *proto.Schema
 	WorkingDir string
 	generator  CodeGenerator
+}
+
+type ScaffoldResult struct {
+	FunctionsCount int
+
+	CreatedFunctions []string
+}
+
+type FunctionImplementation struct {
+	Op    *proto.Operation
+	Model *proto.Model
 }
 
 var SCHEMA_FILE = "schema.keel"
@@ -41,17 +55,74 @@ func NewRuntime(workingDir string) (*Runtime, error) {
 func (r *Runtime) GenerateClient() error {
 	src := r.generator.GenerateClientCode()
 
-	_, err := r.makeModule(filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "index.ts"), src)
+	_, err := r.makeModule(filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "src", "index.ts"), src)
 
-	return err
+	if err != nil {
+		return err
+	}
+
+	return r.GenerateClientTypings()
+}
+
+func (r *Runtime) GenerateClientTypings() error {
+	cmd := exec.Command(
+		"npx",
+		"tsc",
+		"index.ts",
+		"--declaration",
+		"--emitDeclarationOnly",
+		"--declarationDir",
+		"../dist",
+	)
+
+	cmd.Dir = filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "src")
+
+	o, err := cmd.CombinedOutput()
+
+	if err != nil {
+		fmt.Print(string(o))
+		return err
+	}
+
+	return nil
 }
 
 func (r *Runtime) GenerateHandler() error {
 	src := r.generator.GenerateEntryPoint()
 
-	_, err := r.makeModule(filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "handler.ts"), src)
+	_, err := r.makeModule(filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "src", "handler.js"), src)
 
 	return err
+}
+
+//go:embed client-package.json
+var clientPackageJson string
+
+func (r *Runtime) GenerateClientPackageJson() error {
+	packageJsonPath := filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "package.json")
+	clientDir := filepath.Dir(packageJsonPath)
+
+	if _, err := os.Stat(clientDir); errors.Is(err, os.ErrNotExist) {
+		err := os.Mkdir(clientDir, os.ModePerm)
+
+		if err != nil {
+			return err
+		}
+	}
+
+	f, err := os.Create(packageJsonPath)
+
+	if err != nil {
+		return err
+	}
+
+	_, err = f.WriteString(clientPackageJson)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func (r *Runtime) ReconcilePackageJson() error {
@@ -70,7 +141,7 @@ func (r *Runtime) ReconcilePackageJson() error {
 	return nil
 }
 
-// Bundle transpiles all TypeScript in a working directory using
+// Bundle transpiles all generated TypeScript files in a working directory using
 // esbuild, and outputs the JavaScript equivalent to the OutDir
 func (r *Runtime) Bundle(write bool) (api.BuildResult, []error) {
 	// Run esbuild on the generated entrypoint code
@@ -78,15 +149,17 @@ func (r *Runtime) Bundle(write bool) (api.BuildResult, []error) {
 	// so these will be bundled in addition to any generated code
 	buildResult := api.Build(api.BuildOptions{
 		EntryPoints: []string{
-			path.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "index.ts"),
-			path.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "handler.ts"),
+			path.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "src", "index.ts"),
+			path.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "src", "handler.js"),
 		},
-		Bundle:         true,
-		Outdir:         filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "dist"),
-		Write:          write,
-		AllowOverwrite: true,
-		Platform:       api.PlatformNode,
-		LogLevel:       api.LogLevelError,
+		Bundle:   true,
+		Outdir:   filepath.Join(r.WorkingDir, "node_modules", "@teamkeel", "client", "dist"),
+		Write:    write,
+		Platform: api.PlatformNode,
+		LogLevel: api.LogLevelError,
+		External: []string{
+			"@teamkeel/sdk",
+		},
 	})
 
 	if len(buildResult.Errors) > 0 {
@@ -96,34 +169,58 @@ func (r *Runtime) Bundle(write bool) (api.BuildResult, []error) {
 	return buildResult, nil
 }
 
-func (r *Runtime) Scaffold() error {
+func (r *Runtime) Scaffold() (s *ScaffoldResult, e error) {
 	generator := NewCodeGenerator(r.Schema)
 
-	for _, model := range r.Schema.Models {
-		for _, op := range model.Operations {
-			if op.Implementation == proto.OperationImplementation_OPERATION_IMPLEMENTATION_CUSTOM {
-				path := filepath.Join(r.WorkingDir, FUNCTIONS_DIRECTORY, fmt.Sprintf("%s.ts", op.Name))
+	functionsDir := filepath.Join(r.WorkingDir, FUNCTIONS_DIRECTORY)
 
-				if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
-					src := generator.GenerateFunction(model.Name)
+	if _, err := os.Stat(functionsDir); errors.Is(err, os.ErrNotExist) {
+		err := os.Mkdir(functionsDir, os.ModePerm)
 
-					if err != nil {
-						return err
-					}
-
-					err = os.WriteFile(path, []byte(src), 0644)
-
-					fmt.Printf("Scaffolded function %s (%s)", op.Name, path)
-
-					if err != nil {
-						return err
-					}
-				}
-			}
+		if err != nil {
+			return nil, err
 		}
 	}
 
-	return nil
+	funcs := lo.FlatMap(r.Schema.Models, func(m *proto.Model, _ int) (ops []*FunctionImplementation) {
+		for _, op := range m.Operations {
+			if op.Implementation == proto.OperationImplementation_OPERATION_IMPLEMENTATION_CUSTOM {
+				ops = append(ops, &FunctionImplementation{
+					Model: m,
+					Op:    op,
+				})
+			}
+		}
+
+		return ops
+	})
+
+	if len(funcs) == 0 {
+		return &ScaffoldResult{
+			FunctionsCount: 0,
+		}, nil
+	}
+
+	sr := &ScaffoldResult{
+		FunctionsCount: len(funcs),
+	}
+	for _, f := range funcs {
+		path := filepath.Join(r.WorkingDir, FUNCTIONS_DIRECTORY, fmt.Sprintf("%s.ts", f.Op.Name))
+
+		if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
+
+			src := generator.GenerateFunction(f.Op.Name)
+			err = os.WriteFile(path, []byte(src), 0644)
+
+			if err != nil {
+				return sr, err
+			}
+
+			sr.CreatedFunctions = append(sr.CreatedFunctions, path)
+		}
+	}
+
+	return sr, nil
 }
 
 func (r *Runtime) RunServer(port int, onBoot func(process *os.Process)) error {