+++ functions/runtime/build.mjs
@@ -8,7 +8,11 @@ build({
   bundle: true,
   platform: 'node',
   allowOverwrite: true,
-  entryPoints: ['src/index.ts'],
+  entryPoints: [
+    'src/index.ts',
+    'src/analyse.ts',
+    'src/cli.ts'
+  ],
   external: [
     '../client'
   ]